+++ front/app/modules/commercial/insights/hooks/useScanInsightsCategory.test.tsx
@@ -29,6 +29,7 @@ jest.mock(
   () => {
     return {
       insightsTriggerCategoriesSuggestionsTasks: jest.fn(),
+      insightsDeleteCategoriesSuggestionsTasks: jest.fn(),
       insightsCategoriesSuggestionsTasksStream: jest.fn(() => {
         return {
           observable: mockObservable,
@@ -72,6 +73,22 @@ describe('useScanInsightsCategory', () => {
     const { result } = renderHook(() =>
       useScanInsightsCategory(viewId, categoryId)
     );
+
+    await waitFor(() => {
+      expect(result.current.status).toEqual('isIdle');
+    });
+  });
+
+  it('should return correct status when scan is cancelled', async () => {
+    const { result } = renderHook(() =>
+      useScanInsightsCategory(viewId, categoryId)
+    );
+
+    act(() => {
+      result.current.cancelScan();
+    });
+
+    expect(result.current.status).toEqual('isCancelling');
 
     await waitFor(() => {
       expect(result.current.status).toEqual('isIdle');